% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{check_details}
\alias{check_details}
\title{Query R CMD check results and parameters}
\usage{
check_details(check)
}
\arguments{
\item{check}{A check result.}
}
\value{
A named list with elements:
\itemize{
\item \code{package}: package name.
\item \code{version}: package version.
\item \code{rversion}: R version.
\item \code{notes}: character vector of check NOTEs, each NOTE is an
element.
\item \code{warnings}: character vector of check WARNINGs, each WARNING is an
element.
\item \code{errors}: character vector of check ERRORs, each ERROR is an element.
A check timeout adds an ERROR to this vector.
\item \code{platform}: check platform
\item \code{checkdir}: check directory.
\item \code{install_out}: the output of the installation, contents of the
\code{00install.out} file. A single string.
\item \code{description}: the contents of the DESCIRPTION file of the package.
A single string.
\item \code{session_info}: the output of \code{\link[sessioninfo:session_info]{sessioninfo::session_info()}},
from the R session performing the checks.
\item \code{checkdir}: the path to the check directory, if it hasn't been
cleaned up yet, or \code{NA}. The check directory is automatically
cleaned up, when the check object is deleted (garbage collected).
\item \code{cran}: whether it is a CRAN packaged package.
\item \code{bioc}: whether it is a BioConductor package.
}
}
\description{
Query R CMD check results and parameters
}
