% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_checks}
\alias{compare_checks}
\title{Compare a set of check results to another check result}
\usage{
compare_checks(old, new)
}
\arguments{
\item{old}{A check result, or a list of check results.}

\item{new}{A check result.}
}
\value{
An \code{rcmdcheck_comparison} object with fields:
\itemize{
\item \code{package}: the name of the package, string,
\item \code{versions}: package versions, length two character,
\item \code{status}: comparison status, see below,
\item \code{old}: list of \code{rcmdcheck} objects the old check(s),
\item \code{new}: \code{rcmdcheck} object, the new check,
\item \code{cmp}:
}
}
\description{
Compare a set of check results to another check result
}
\seealso{
Other check comparisons:
\code{\link[=compare_to_cran]{compare_to_cran()}}
}
\concept{check comparisons}
