% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{cran_check_flavours}
\alias{cran_check_flavours}
\title{Download and show all CRAN check flavour platforms}
\usage{
cran_check_flavours(package = NULL)
}
\arguments{
\item{package}{CRAN package name or \code{NULL}.}
}
\value{
Character vector of platform ids.
}
\description{
If the \code{package} argument is \code{NULL}, then all current
platforms are downloaded. If the \code{package} argument is specified,
then all flavours used for the latests package checks for that
package, are downloaded and returned.
}
\examples{
\dontrun{
cran_check_flavours()
cran_check_flavours("simplegraph")
}
}

