% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{cran_check_results}
\alias{cran_check_results}
\title{Download and parse R CMD check results from CRAN}
\usage{
cran_check_results(package, flavours = cran_check_flavours(package))
}
\arguments{
\item{package}{Name of a single package to download the checks for.}

\item{flavours}{CRAN check flavours to use. Defaults to all
flavours that were used to check the package.}
}
\value{
A list of \code{rcmdcheck} objects.
}
\description{
Download and parse R CMD check results from CRAN
}
