% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_frame}
\alias{data_frame}
\title{Alternative to data.frame}
\usage{
data_frame(..., stringsAsFactors = FALSE)
}
\arguments{
\item{...}{Named data frame columns, or data frames.}

\item{stringsAsFactors}{Just leave it on \code{FALSE}. \code{:)}}
}
\value{
Data frame.
}
\description{
\itemize{
\item Sets stringsAsFactors to FALSE by default
\item If any columns have zero length, the result will have
zero rows.
\item If a column is a scalar, then it will be recycled.
\item Non-matching number of rows gives an error, except for
lengths zero and one.
}
}
\keyword{internal}
