% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myrep}
\alias{myrep}
\title{Recycle a vector or a data frame (rows)}
\usage{
myrep(x, len)
}
\arguments{
\item{x}{Vector or data frame to replicate. Must be length 0, 1, or
len.}

\item{len}{Expected length.}
}
\description{
Recycle a vector or a data frame (rows)
}
\keyword{internal}
