% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_check}
\alias{parse_check}
\title{Parse \code{R CMD check} results from a file or string}
\usage{
parse_check(file = NULL, text = NULL)
}
\arguments{
\item{file}{The \code{00check.log} file, or a directory that
contains that file. It can also be a connection object.}

\item{text}{The contentst of a \code{00check.log} file.}
}
\value{
An \code{rcmdcheck} object, the check results.
}
\description{
At most one of \code{file} or \code{text} can be given.
If both are \code{NULL}, then the current working directory
is checked for a \code{00check.log} file.
}
\seealso{
\code{\link{parse_check_url}}
}
