% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_check_url}
\alias{parse_check_url}
\title{Shorthand to parse R CMD check results from a URL}
\usage{
parse_check_url(url, quiet = TRUE)
}
\arguments{
\item{url}{URL to parse the results from. Note that it should
not contain HTML markup, just the text output.}

\item{quiet}{Passed to \code{download.file}.}
}
\value{
An \code{rcmdcheck} object, the check results.
}
\description{
Shorthand to parse R CMD check results from a URL
}
\seealso{
\code{\link{parse_check}}
}

