% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rcmdcheck}
\alias{print.rcmdcheck}
\title{Print R CMD check results}
\usage{
\method{print}{rcmdcheck}(x, headers = TRUE, ...)
}
\arguments{
\item{header}{Whether to print a header.}
}
\description{
Print R CMD check results
}

