% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{print.rcmdcheck_comparison}
\alias{print.rcmdcheck_comparison}
\title{Print R CMD check result comparisons}
\usage{
\method{print}{rcmdcheck_comparison}(x, header = TRUE, ...)
}
\arguments{
\item{x}{R CMD check result comparison object.}

\item{header}{Whether to print the header. You can suppress the
header if you want to use the printout as part of another object's
printout.}

\item{...}{Additional arguments, currently ignored.}
}
\description{
See \code{\link[=compare_checks]{compare_checks()}} and \code{\link[=compare_to_cran]{compare_to_cran()}}.
}
