% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmdcheck-package.R
\docType{package}
\name{rcmdcheck-package}
\alias{rcmdcheck-package}
\title{rcmdcheck: Run 'R CMD check' from 'R' and Capture Results}
\description{
Run 'R CMD check' from 'R' and capture the results of the individual checks. Supports running checks in the background, timeouts, pretty printing and comparing check results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rcmdcheck.r-lib.org/}
  \item \url{https://github.com/r-lib/rcmdcheck#readme}
  \item Report bugs at \url{https://github.com/r-Lib/rcmdcheck/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Gábor Csárdi \email{csardi.gabor@gmail.com}
}

Other contributors:
\itemize{
  \item Mango Solutions [copyright holder, funder]
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
