% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{rcmdcheck_bg}
\alias{rcmdcheck_bg}
\title{Start running \code{R CMD check} in the background}
\usage{
rcmdcheck_bg(path = ".", args = character(), libpath = .libPaths(),
  repos = getOption("repos"))
}
\arguments{
\item{path}{Path to a package tarball or a directory.}

\item{args}{Character vector of arguments to pass to
\code{R CMD check}.}

\item{libpath}{The library path to set for the check.
The default uses the current library path.}

\item{repos}{The \code{repos} option to set for the check.
This is needed for cyclic dependency checks if you use the
\code{--as-cran} argument. The default uses the current value.}
}
\value{
A list with two elements:
\itemize{
\item \code{process} is the \link[processx:process]{processx::process} object of the background process.
\item \code{targz} is the (temporary) \code{tar.gz} package file.
}
}
\description{
Start running \code{R CMD check} in the background
}
