% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xopen.R
\name{xopen.rcmdcheck}
\alias{xopen.rcmdcheck}
\title{Open the check directory in a file browser window}
\usage{
\method{xopen}{rcmdcheck}(target, app = NULL, quiet = FALSE, ...)
}
\arguments{
\item{target}{\code{rcmdcheck()} result.}

\item{app}{Specify the app to open \code{target} with, and its arguments,
in a character vector. Note that app names are platform dependent.}

\item{quiet}{Whether to echo the command to the screen, before
running it.}

\item{...}{Additional arguments, not used currently.}
}
\description{
Open the check directory in a file browser window
}
