% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_checks}
\alias{compare_checks}
\title{Compare a set of check results to another check result}
\usage{
compare_checks(old, new)
}
\arguments{
\item{old}{A check result, or a list of check results.}

\item{new}{A check result.}
}
\value{
An \code{rcmdcheck_comparison} object.
}
\description{
Compare a set of check results to another check result
}
\seealso{
Other check comparisons: \code{\link{compare_to_cran}}
}
\concept{check comparisons}
