% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_to_cran}
\alias{compare_to_cran}
\title{Compare a check result to CRAN check results}
\usage{
compare_to_cran(check, flavours = cran_check_flavours(check$package))
}
\arguments{
\item{check}{A check result.}

\item{flavours}{CRAN check flavour(s) to use. By default all
platforms are used.}
}
\value{
An code{rmdcheck_comparison} object.
}
\description{
Compare a check result to CRAN check results
}
