% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rcmdcheck}
\alias{print.rcmdcheck}
\title{Print R CMD check results}
\usage{
\method{print}{rcmdcheck}(
  x,
  header = TRUE,
  test_output = getOption("rcmdcheck.test_output", FALSE),
  ...
)
}
\arguments{
\item{x}{Check result object to print.}

\item{header}{Whether to print a header.}

\item{test_output}{if \code{TRUE}, include the test output in the results, even
if there are no test failures.}

\item{...}{Additional arguments, currently ignored.}
}
\description{
Print R CMD check results
}
