% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rcmdcheck}
\alias{print.rcmdcheck}
\title{Print R CMD check results}
\usage{
\method{print}{rcmdcheck}(x, header = TRUE, ...)
}
\arguments{
\item{x}{Check result object to print.}

\item{header}{Whether to print a header.}

\item{...}{Additional arguments, currently ignored.}
}
\description{
Print R CMD check results
}

