% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rcmdcheck-config}
\alias{rcmdcheck-config}
\title{rcmdcheck configuration}
\description{
Options take precedence over environment variables. E.g. if both
the \code{RCMDCHECK_NUM_COLORS} environment variables and the
\code{rcmdcheck.num_colors} option are set, then the latter is used.
}
\details{
rcmdcheck uses the cli package for much of its output, so you can
configure the output via cli, see \link[cli:cli-config]{cli::cli-config}.
}
\section{Environment variables}{
\itemize{
\item \code{R_PROFILE_USER}: standard R environment variable to configure the
path to the user level R profile. See \link[base:Startup]{base::R_PROFILE_USER}.
\item \code{RCMDCHECK_BASE_URL}: URL to the root of the CRAN check web page.
You can use this to select an alternative CRAN mirror. Defaults to
\verb{https://cran.r-project.org/web/checks/}.
\item \code{RCMDCHECK_DETAILS_URL}: URL to the root of the CRAN check output
page. Defaults to \verb{https://www.r-project.org/nosvn/R.check/}.
\item \code{RCMDCHECK_FLAVOURS_URL} URL to the CRAN check flavours page.
You can use this to select an alternative CRAN mirror. Defaults to
\verb{https://cran.r-project.org/web/checks/check_flavors.html}.
\item \code{RCMDCHECK_NUM_COLORS}: the number of ANSI colors to use in the output.
It can be used to override the number of colors detected or configured
by the cli package. See \code{\link[cli:num_ansi_colors]{cli::num_ansi_colors()}}. This configuration
is only used for the output of rcmdcheck and it does not affect the
examples and test cases (and other code) of the checked package.
It not set, then the default of cli is uesed. The corresponding
option is \code{rcmdcheck.num_colors}.
\item \code{RCMDCHECK_TIMESTAMP_LIMIT}: lower limit is seconds, above which
rcmdcheck adds time stamps to the individual check steps. It may be
fractional. Defaults to 1/3 of a second.
}
}

\section{Options}{
\itemize{
\item \code{rcmdcheck.num_colors}
}
}

