% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rcmdcheck}
\alias{rcmdcheck}
\alias{rcmdcheck-package}
\title{Run R CMD check from R and Capture Results}
\description{
Run R CMD check form R programatically, and capture the results of the
individual checks.
}

